Moved wait() into Navigation
- moved sleep() into helpers/navigation
- deleted helpers/utilties
- consolidated tests (tests related were moved into
  */navigation.spec.ts)
- Resolve issue #51
