+++ src/botmation/helpers/navigation.ts
@@ -9,3 +9,10 @@ export const enrichGoToPageOptions = (overloadDefaultOptions: Partial<DirectNavi
   waitUntil: 'networkidle0',
   ...overloadDefaultOptions
 })
+
+/**
+ * @description   Delay execution of next line by X milliseconds, promisified so you can await a setTimeout call
+ * @param milliseconds 
+ */
+export const sleep = async(milliseconds: number): Promise<NodeJS.Timeout> =>
+  new Promise(resolve => setTimeout(resolve, milliseconds))
